govt-schemes-assistant/requirements.txt
streamlit
pyyaml
huggingface_hub
python-dotenv
